positive

+++ 0-positive_or_negative.c
@@ -1,6 +1,6 @@
 #include <stdlib.h>
-#include <time.h>
 #include <stdio.h>
+#include <time.h>
 /**
  * main - Determines if a number is positive, negative or zero.
  *
@@ -11,7 +11,7 @@ int main(void)
 	int n;
 
 	srand(time(0));
-	n = rand() - RAND_MAX / 2;
+	n = rand() - (int) RAND_MAX / 2;
 	if (n > 0)
 	{
 		printf("%d is positive\n", n);
